pyproject.toml
[project]
name = "rnn-sequence-classification-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Hlib Khorunzhyi"}
]
requires-python = "~=3.12"
dependencies = [
    "hydra-core~=1.3.2",
    "pydantic~=2.13.4",
    "google-cloud-secret-manager>=2.16.1",
    "dvc[gdrive,gs]>=2.56.0",
    "pandas>=2.0",
    "nltk>=3.8",
    "symspellpy~=6.9",
    "fsspec[gcs]>=2023.5",
    "gcsfs>=2023.4",
    "fastparquet~=2026.3",
    "distributed>=2023.5",
    "dask-cloudprovider[gcp]>=2022.10",
    "dask-ml>=2023.3"
]

[tool.poetry.group.dev.dependencies]
jupyterlab = ">=3.6.3"
pytest = ">=7.3"
black = ">=23.3"
isort = ">5.12"
flake8 = ">=6.0"
mypy = ">=1.2"

types-pyYAML = "~=5.4.7"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]
skip-string-normalization = false
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["src"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
force_grid_wrap = 0
combine_as_imports = true

[tool.mypy]
python_version = 3.12
ignore_missing_imports = true
strict = true
disallow_untyped_defs = true
warn_unused_ignores = true
show_error_codes = true
